requirements.txt
# Core dependencies
streamlit>=1.22.0
python-dotenv>=1.0.0

# LangChain and related components
langchain>=0.0.267
langchain-google-genai>=0.0.3
langchain-community>=0.0.10
langchain-core>=0.0.16

# Vector database and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Document processing
pypdf>=3.15.1
unstructured>=0.10.12
tabulate>=0.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0

# Database
sqlalchemy>=2.0.0